server/requirements.txt
bleach==6.2.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
Flask==3.1.2
flask-cors==6.0.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kaggle==1.7.4.5
kagglehub==0.3.13
MarkupSafe==3.0.2
numpy==2.3.2
packaging==25.0
pluggy==1.6.0
protobuf==6.32.0
Pygments==2.19.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.5
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
text-unidecode==1.3
threadpoolctl==3.6.0
tqdm==4.67.1
urllib3==2.5.0
webencodings==0.5.1
Werkzeug==3.1.3
